MarcXmlMapping trait
